# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HDP App OpenApi API Documentation",
        "description": "HDP App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "hdp@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customer/consultations": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Show all consultations of a customer",
                "description": "Show all consultations of a customer",
                "operationId": "allFacilities",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/customer/customer-depandants": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Show all facilities ",
                "description": "Show all facilities ",
                "operationId": "allFacilities",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "customer"
                ],
                "summary": "create a customer's dependant ",
                "description": "create a customer's dependant ",
                "operationId": "createCustomerDependant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerDependantCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/customer/customer-depandants/{id}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "show a customer's dependant ",
                "description": "show a customer's dependant ",
                "operationId": "showCustomerDependant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "customer"
                ],
                "summary": "update a customer's dependant ",
                "description": "update a customer's dependant ",
                "operationId": "updateCustomerDependant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerDependantUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "customer"
                ],
                "summary": "delete a customer's dependant ",
                "description": "delete a customer's dependant ",
                "operationId": "deleteCustomerDependant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dependant ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/customer/customer-service-requests/{id}": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "show a Service requests to customer ",
                "description": "show a Service requests to customer ",
                "operationId": "showCustomerDependant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/general/general-facilities": {
            "get": {
                "tags": [
                    "general"
                ],
                "summary": "Show all facilities ",
                "description": "Show all facilities ",
                "operationId": "allFacilities",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/professional/professional-service-availability": {
            "get": {
                "tags": [
                    "professional"
                ],
                "summary": "Show all service availabilities of a professional",
                "description": "Show all service availabilities of a professional",
                "operationId": "allFacilities",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "professional"
                ],
                "summary": "create a professional's dependant ",
                "description": "create a professional's dependant ",
                "operationId": "createProfessionalServiceAvailability",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfessionalServiceAvailabilityCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/professional/professional-service-offerings": {
            "get": {
                "tags": [
                    "professional"
                ],
                "summary": "Show all service offerings of a professional",
                "description": "Show all service offerings of a professional",
                "operationId": "allFacilities",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "professional"
                ],
                "summary": "create a service offerings ",
                "description": "create a service offerings ",
                "operationId": "createProfessionalServiceOffering",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfessionalServiceOfferingCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomerDependantCreateFormRequest": {
                "title": "Customer Dependant Create Form Request Fields",
                "description": "Customer Dependant Create request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "dependant_name": {
                        "title": " Offer Name",
                        "description": " Offer Name",
                        "type": "string",
                        "example": "name"
                    },
                    "telephone": {
                        "title": "Telephone",
                        "description": "Telephone",
                        "type": "string",
                        "example": "070384444444"
                    },
                    "birth_date": {
                        "title": "Birth Date",
                        "description": "Birth Date",
                        "type": "string",
                        "example": "28-04-1991"
                    }
                },
                "type": "object"
            },
            "CustomerDependantUpdateFormRequest": {
                "title": "Customer Dependant Update Form Request Fields",
                "description": "Customer Dependant Update request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "dependant_name": {
                        "title": " Offer Name",
                        "description": " Offer Name",
                        "type": "string",
                        "example": "name"
                    },
                    "telephone": {
                        "title": "Telephone",
                        "description": "Telephone",
                        "type": "string",
                        "example": "070384444444"
                    },
                    "birth_date": {
                        "title": "Birth Date",
                        "description": "Birth Date",
                        "type": "string",
                        "example": "28-04-1991"
                    }
                },
                "type": "object"
            },
...
            "ProfessionalServiceAvailabilityCreateFormRequest": {
                "title": "Professional Service Availability Create Form Request Fields",
                "description": "Professional Service Availability Create request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "professional_id": {
                        "title": "Professional Id",
                        "description": "Professional Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "service_offering_id": {
                        "title": "Service Offering Id",
                        "description": "Service Offering Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "is_available_monday": {
                        "title": "Is professional Available on Monday",
                        "description": "Is professional Available on Monday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_tuesday": {
                        "title": "Is professional Available on tuesday",
                        "description": "Is professional Available on tuesday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_wednesday": {
                        "title": "Is professional Available on wednesday",
                        "description": "Is professional Available on wednesday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_thursday": {
                        "title": "Is professional Available on thursday",
                        "description": "Is professional Available on thursday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_friday": {
                        "title": "Is professional Available on friday",
                        "description": "Is professional Available on friday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_saturday": {
                        "title": "Is professional Available on saturday",
                        "description": "Is professional Available on saturday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "is_available_sunday": {
                        "title": "Is professional Available on sunday",
                        "description": "Is professional Available on sunday",
                        "type": "boolean",
                        "example": "true/false"
                    },
                    "specific_time_open": {
                        "title": "Is professional Open time",
                        "description": "Is professional Open time",
                        "type": "string",
                        "example": "9:00"
                    },
                    "specific_time_close": {
                        "title": "Is professional Close time",
                        "description": "Is professional Close time",
                        "type": "string",
                        "example": "22:00"
                    },
                    "is_currently_unavailable": {
                        "title": "Is professional Currently Available",
                        "description": "Is professional Currently Available",
                        "type": "boolean",
                        "example": "true/false"
                    }
                },
                "type": "object"
            },
...
            "ProfessionalServiceOfferingCreateFormRequest": {
                "title": "Professional Service Offering Create Form Request Fields",
                "description": "Professional Service Offering Create request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "professional_id": {
                        "title": "Professional Id",
                        "description": "Professional Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "service_offering_id": {
                        "title": "Service Offering Id",
                        "description": "Service Offering Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "offer_name": {
                        "title": " Offer Name",
                        "description": " Offer Name",
                        "type": "string",
                        "example": "name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "Description"
                    },
                    "delivery_mode": {
                        "title": "Delivery mode",
                        "description": "Delivery mode",
                        "type": "string",
                        "example": "Delivery mode"
                    },
                    "profile_image": {
                        "title": "Profile  image",
                        "description": "Profile  image",
                        "type": "number",
                        "format": "float",
                        "example": "Profile  image"
                    },
                    "offer_unit_price": {
                        "title": "Offer Unit Price",
                        "description": "Offer Unit Price",
                        "type": "number",
                        "format": "float",
                        "example": "Offer Unit Price"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}